+++ src/Symfony/Component/PropertyInfo/Type.php
@@ -57,9 +57,12 @@ class Type
     private $collectionValueType;
 
     /**
+     * @param Type[]|Type|null $collectionKeyType
+     * @param Type[]|Type|null $collectionValueType
+     *
      * @throws \InvalidArgumentException
      */
-    public function __construct(string $builtinType, bool $nullable = false, string $class = null, bool $collection = false, self $collectionKeyType = null, self $collectionValueType = null)
+    public function __construct(string $builtinType, bool $nullable = false, string $class = null, bool $collection = false, $collectionKeyType = null, $collectionValueType = null)
     {
         if (!\in_array($builtinType, self::$builtinTypes)) {
             throw new \InvalidArgumentException(sprintf('"%s" is not a valid PHP type.', $builtinType));
@@ -69,8 +72,31 @@ class Type
         $this->nullable = $nullable;
         $this->class = $class;
         $this->collection = $collection;
-        $this->collectionKeyType = $collectionKeyType;
-        $this->collectionValueType = $collectionValueType;
+        $this->collectionKeyType = $this->validateCollectionArgument($collectionKeyType, 5, '$collectionKeyType') ?? [];
+        $this->collectionValueType = $this->validateCollectionArgument($collectionValueType, 6, '$collectionValueType') ?? [];
+    }
+
+    private function validateCollectionArgument($collectionArgument, int $argumentIndex, string $argumentName): ?array
+    {
+        if (null === $collectionArgument) {
+            return null;
+        }
+
+        if (!\is_array($collectionArgument) && !$collectionArgument instanceof self) {
+            throw new \TypeError(sprintf('"%s()": Argument #%d (%s) must be of type "%s[]", "%s" or "null", "%s" given.', __METHOD__, $argumentIndex, $argumentName, self::class, self::class, get_debug_type($collectionArgument)));
+        }
+
+        if (\is_array($collectionArgument)) {
+            foreach ($collectionArgument as $type) {
+                if (!$type instanceof self) {
+                    throw new \TypeError(sprintf('"%s()": Argument #%d (%s) must be of type "%s[]", "%s" or "null", array value "%s" given.', __METHOD__, $argumentIndex, $argumentName, self::class, self::class, get_debug_type($collectionArgument)));
+                }
+            }
+
+            return $collectionArgument;
+        }
+
+        return [$collectionArgument];
     }
 
     /**
@@ -107,8 +133,33 @@ class Type
      * Gets collection key type.
      *
      * Only applicable for a collection type.
+     *
+     * @deprecated since Symfony 5.3, use "getCollectionKeyTypes()" instead
      */
     public function getCollectionKeyType(): ?self
+    {
+        trigger_deprecation('symfony/property-info', '5.3', 'The "%s()" method is deprecated, use "getCollectionKeyTypes()" instead.', __METHOD__);
+
+        $type = $this->getCollectionKeyTypes();
+        if (0 === \count($type)) {
+            return null;
+        }
+
+        if (\is_array($type)) {
+            [$type] = $type;
+        }
+
+        return $type;
+    }
+
+    /**
+     * Gets collection key types.
+     *
+     * Only applicable for a collection type.
+     *
+     * @return Type[]
+     */
+    public function getCollectionKeyTypes(): array
     {
         return $this->collectionKeyType;
     }
@@ -117,8 +168,33 @@ class Type
      * Gets collection value type.
      *
      * Only applicable for a collection type.
+     *
+     * @deprecated since Symfony 5.3, use "getCollectionValueTypes()" instead
      */
     public function getCollectionValueType(): ?self
+    {
+        trigger_deprecation('symfony/property-info', '5.3', 'The "%s()" method is deprecated, use "getCollectionValueTypes()" instead.', __METHOD__);
+
+        $type = $this->getCollectionValueTypes();
+        if (0 === \count($type)) {
+            return null;
+        }
+
+        if (\is_array($type)) {
+            [$type] = $type;
+        }
+
+        return $type;
+    }
+
+    /**
+     * Gets collection value types.
+     *
+     * Only applicable for a collection type.
+     *
+     * @return Type[]
+     */
+    public function getCollectionValueTypes(): array
     {
         return $this->collectionValueType;
     }